+++ app/Models/Anotacion.php
@@ -2,12 +2,15 @@
 
 namespace App\Models;
 
+use App\Traits\RegistrableCambio;
+
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 class Anotacion extends Model
 {
+use RegistrableCambio;
     use HasFactory;
 
     protected $table = 'anotaciones';