+++ app/Models/User.php
@@ -2,6 +2,8 @@
 
 namespace App\Models;
 
+use App\Traits\RegistrableCambio;
+
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
@@ -9,6 +11,7 @@ use Laravel\Sanctum\HasApiTokens;
 
 class User extends Authenticatable
 {
+use RegistrableCambio;
     use HasApiTokens, HasFactory, Notifiable;
 
     /**